Move all recursive validation tests into one file
The `/tests` folder bloated up a little bit with all these
recursive validation tests hence merge all of them into one
file with modules seperating them.
